Add FreeBSD support to ZoL
- Refactor tree in to os specific parts
- Import FreeBSD SPL
- update tests to work on FreeBSD

+++ tests/zfs-tests/tests/functional/slog/slog_replay_volume.ksh
@@ -64,7 +64,7 @@ MNTPNT=$TESTDIR/$TESTVOL
 
 function cleanup_volume
 {
-	if ismounted $MNTPNT ext4; then
+	if ismounted $MNTPNT $NEWFS_DEFAULT_FS; then
 		log_must umount $MNTPNT
 		rmdir $MNTPNT
 	fi
@@ -88,10 +88,16 @@ log_must zfs set compression=on $TESTPOOL/$TESTVOL
 log_must zfs set sync=always $TESTPOOL/$TESTVOL
 log_must mkdir -p $TESTDIR
 block_device_wait
-echo "y" | newfs -t ext4 -v $VOLUME
-log_must mkdir -p $MNTPNT
-log_must mount -o discard $VOLUME $MNTPNT
-log_must rmdir $MNTPNT/lost+found
+if is_freebsd; then
+	log_must newfs $VOLUME
+	log_must mkdir -p $MNTPNT
+	log_must mount $VOLUME $MNTPNT
+else
+	log_must eval "echo y | newfs -t ext4 -v $VOLUME"
+	log_must mkdir -p $MNTPNT
+	log_must mount -o discard $VOLUME $MNTPNT
+	log_must rmdir $MNTPNT/lost+found
+fi
 log_must zpool sync
 
 #
@@ -104,8 +110,13 @@ log_must zpool freeze $TESTPOOL
 #
 
 # TX_WRITE
-log_must dd if=/dev/urandom of=$MNTPNT/latency-8k bs=8k count=1 oflag=sync
-log_must dd if=/dev/urandom of=$MNTPNT/latency-128k bs=128k count=1 oflag=sync
+if is_freebsd; then
+	log_must dd if=/dev/urandom of=$MNTPNT/latency-8k bs=8k count=1 conv=sync
+	log_must dd if=/dev/urandom of=$MNTPNT/latency-128k bs=128k count=1 conv=sync
+else
+	log_must dd if=/dev/urandom of=$MNTPNT/latency-8k bs=8k count=1 oflag=sync
+	log_must dd if=/dev/urandom of=$MNTPNT/latency-128k bs=128k count=1 oflag=sync
+fi
 
 # TX_WRITE (WR_INDIRECT)
 log_must zfs set logbias=throughput $TESTPOOL/$TESTVOL
@@ -128,7 +139,11 @@ fi
 #
 # 4. Generate checksums for all ext4 files.
 #
-typeset checksum=$(cat $MNTPNT/* | sha256digest)
+if is_freebsd; then
+	typeset checksum=$(cat $MNTPNT/* | sha256)
+else
+	typeset checksum=$(cat $MNTPNT/* | sha256digest)
+fi
 
 #
 # 5. Unmount filesystem and export the pool
@@ -160,7 +175,11 @@ log_note "Verify current block usage:"
 log_must zdb -bcv $TESTPOOL
 
 log_note "Verify checksums"
-typeset checksum1=$(cat $MNTPNT/* | sha256digest)
+if is_freebsd; then
+	typeset checksum1=$(cat $MNTPNT/* | sha256)
+else
+	typeset checksum1=$(cat $MNTPNT/* | sha256digest)
+fi
 [[ "$checksum1" == "$checksum" ]] || \
     log_fail "checksum mismatch ($checksum1 != $checksum)"
